+++ tramex-tools/src/interface/interface_file/mod.rs
@@ -2,6 +2,3 @@
 
 pub mod file_handler;
 pub mod utils_file;
-
-pub mod parser;
-pub mod parser_rrc;
